GTK: Fix bookmark manager crash on startup introduced by r20565
TBR=munjal


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20899 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/gtk/bookmark_manager_gtk.h
@@ -10,6 +10,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include <vector>
 
 #include "app/table_model_observer.h"
+#include "base/basictypes.h"
 #include "base/ref_counted.h"
 #include "base/task.h"
 #include "chrome/browser/bookmarks/bookmark_model.h"
@@ -87,7 +88,7 @@ class BookmarkManagerGtk : public BookmarkModelObserver,
   void BuildRightStore();
 
   // Get the ID of the item at |iter|.
-  int GetRowIDAt(GtkTreeModel* model, GtkTreeIter* iter);
+  int64 GetRowIDAt(GtkTreeModel* model, GtkTreeIter* iter);
 
   // Get the node from |model| at |iter|. If the item is not a node, return
   // NULL.
@@ -141,7 +142,7 @@ class BookmarkManagerGtk : public BookmarkModelObserver,
   // Tries to find the node with id |target_id|. If found, returns true and set
   // |iter| to point to the entry. If you pass a |iter| with stamp of 0, then it
   // will be treated as the first row of |model|.
-  bool RecursiveFind(GtkTreeModel* model, GtkTreeIter* iter, int target_id);
+  bool RecursiveFind(GtkTreeModel* model, GtkTreeIter* iter, int64 target_id);
 
   // Search helpers.
   void PerformSearch();
